+++ core/structures.rs
@@ -1,0 +1,5 @@
+mod slang_dependency;
+mod slang;
+
+pub use slang_dependency::SlangDependency;
+pub use slang::Slang;