+++ core/utils/restricted_names.rs
@@ -1,0 +1,20 @@
+use std::path::Path;
+
+/// These names cannot be used on Windows, even with an extension.
+pub fn is_windows_reserved(name: &str) -> bool {
+    [
+        "con", "prn", "aux", "nul", "com1", "com2", "com3", "com4", "com5", "com6", "com7", "com8",
+        "com9", "lpt1", "lpt2", "lpt3", "lpt4", "lpt5", "lpt6", "lpt7", "lpt8", "lpt9",
+    ]
+    .contains(&name.to_ascii_lowercase().as_str())
+}
+
+/// Check the entire path for names reserved in Windows.
+pub fn is_windows_reserved_path(path: &Path) -> bool {
+    path.iter()
+        .filter_map(|component| component.to_str())
+        .any(|component| {
+            let stem = component.split('.').next().unwrap();
+            is_windows_reserved(stem)
+        })
+}