+++ parser/lib.rs
@@ -22,12 +22,6 @@ pub fn run_file(file: SimpleFile<String, String>) -> Option<Vec<Tok>> {
     }
 }
 
-pub fn run_test() {
-    println!(
-        "Tokens: {:?}",
-        run_file(SimpleFile::new(
-            "test.sf".to_string(),
-            "print('Hello world!')".to_string(),
-        ))
-    );
+pub fn run_content(file_name: String, file_content: String) -> Option<Vec<Tok>> {
+    run_file(SimpleFile::new(file_name, file_content))
 }