+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
 ## 7/11/2020
+- Add `hello_world` and `fibonacci` examples.
+- Add `init` and `start` commands.
 - Add lexer test.
 - Add the lexer and tokens.